backend/requirements.txt
flask
flask-cors
gevent
joblib
mongoengine
openpyxl
pandas
pyjwt
pymongo
python-dotenv
torch
transformers
werkzeug